Add files via upload
Hello and welcome to this course outline my name is Awal Manga and in this course outline  we are going to study mern stack. So mern stack is actually very interesting concept in the world of web development.
MERN STACK is actually a frame work or a set of frame work that we use for web development where each of these alphabets M,E,R,N stand for there own technologies namingly Mongodb, E for Express, R for react and N for node.
 So these four technologies will make you eligible for being a full stack web developer and that is exactly what we are going to learn in this course of line.
MERN STACK as we have discussed is gonna be something  that is gonna be around full stack web development there are quite a few things that we need to understand a full stack web developmer is a person who knows quite a few things that we need to understand.
A full stack web developer is a person who knows quite a few technologies front end and back end development and so on and on with the combination of expressing order to make our data persistent we make use of mongodb.
So this is the whole agenda of this course and we are going to cover around in this entire course outline. Lets move ahead and start this course by installing the right software and tools that will required in order to start this process so lets go ahead and get started.
So the first thing that i am gonna do is open my browser which is safari and after openeing my browser what i will do is basically move ahead and to understanding to what kind of software that i would needing so the fisrt software that i would need is gonna be this thing.
Let me just move the browser so the first software that i would need is gonna be this thing. This is my browser alright my browser is here and the first thing that i am gonna use to download is gonna be node js. So as soon as you right node js you will come to a website call as node js.org/en.
And in this particuar website we will start to see the versions of node js that is present on this particular website as of now i am using mac os because of which i am seeing the available versions of macos.
In case you are using windows or linux you will see macos or node js versions for the right operating system that we are using if i move to the dowload section we will see and choose right installer for you if i move back to the previous screen
we will see that there are two options available first opetion is 18.16.0 LTS and the other option is 20.2.0 Current. This is the current situation of node js as of 29th of may 2023 indian standard time.
By the time your are watching this video the versions of node js might be different but even in that scenario we will see two types of versions of node js LTS and current. LTS stands for long term support and Current stands for latest upgradation of node js.
I will always suggest you to follow the LTS versions and to install in the machine after installing node js on the machine and open terminal and after opening terminal.
what we will do is we can check if the nodejs version has been install on the machine or not so lets move and check if that is correctly present on our machine or not.
So this is my terminal which i have opened over here and i in order to check if it installed on my machine i will write node --version or i can just write -v. 
Press enter you will see that 18.15.0 is the version of nodejs that is installed on my machine right it is currently installed on my machine in your case it will be 16.0. This is gonna be the version that is gonna pop up inside your machine.
The next step that you should be doing is by writing npm -v press enter and it will show you the right version of npm which is installed on your machine in my case it is 9.5.0 make sure that you run these commands after you are done installing the node js on your machine.
The next thing that i am gonna do is i will open safari again and i will download visual studio code so this gonna be the idb we are going to use on visual studio code 
So this gonna be the IDE that we are going to be use in order to start our visual studio code or in order to start our java script development again i am using macos so i will be downloading the right version for macos you can choose to download the right version for your own operating system.
In my own system we can see that visual studio code is already downloaded so I can just open it and start to program, in your case you can download it in your own machine so this is pretty much it for this video where we download the right software .
In the next video we will try to focus on understanding more important stuff about javascript and trying to understand what are the other prerequists that we will be using for this course untill then stay connected to the cipherschool.